Hello from credbot!
Summary:
This is a commit to master under the name and email of credbot, who has
no other contributions to the repository. This is intended to test that
we can still pull the correct GitHub user off of the commit.

Generated with:

    git -c user.name='credbot' \
        -c user.email='42819382+credbot@users.noreply.github.com' \
        commit

Actually committed and signed by William Chargin <wchargin@gmail.com>.
Verify public key at either of:
  - <https://github.com/wchargin.gpg>
  - <https://wchargin.github.io/> (click link to "My PGP key")